chore: share interactive example component
