+++ data/git-timestamps-modified.json
@@ -487,7 +487,7 @@
   "src/pages/index.js": "2026-04-28T15:02:07.000Z",
   "src/pages/insights.js": "2026-04-29T08:52:10.000Z",
   "src/pages/integrations.js": "2020-07-29T16:14:34.000Z",
-  "src/pages/link-preview.js": "2026-05-19T19:40:52.000Z",
+  "src/pages/link-preview.js": "2026-05-20T08:46:43.000Z",
   "src/pages/logo.js": "2026-04-29T09:02:20.000Z",
   "src/pages/markdown.js": "2026-05-02T09:12:58.000Z",
   "src/pages/meta.js": "2026-01-24T13:05:07.000Z",
