+++ src/pages/alternative/embedly.js
@@ -29,7 +29,7 @@ import { Link } from 'components/elements/Link'
 import Meta from 'components/elements/Meta/Meta'
 import SubheadBase from 'components/elements/Subhead'
 import Text from 'components/elements/Text'
-import { InteractiveExample } from 'pages/embed'
+import { InteractiveExample } from 'components/pages/embed'
 
 const Heading = withTitle(HeadingBase)
 const Subhead = withTitle(SubheadBase)
@@ -829,7 +829,8 @@ const Hero = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'embedly' })}
+            trackEvent('alternative cta', { competitor: 'embedly' })
+          }
         >
           Get Started Free
         </ArrowLink>
@@ -1207,7 +1208,8 @@ const CTASection = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'embedly' })}
+            trackEvent('alternative cta', { competitor: 'embedly' })
+          }
           css={theme({ fontSize: 3, px: 5, py: 3, color: 'white' })}
         >
           Start Building Free
@@ -1386,7 +1388,8 @@ const TryItSection = () => (
         <Link
           href='/docs/guides/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'embedly' })}
+            trackEvent('alternative cta', { competitor: 'embedly' })
+          }
           css={theme({ fontSize: CTA_LINK_FONT_SIZE })}
         >
           Start now for free
