+++ src/pages/alternative/iframely.js
@@ -29,7 +29,7 @@ import { Link } from 'components/elements/Link'
 import Meta from 'components/elements/Meta/Meta'
 import SubheadBase from 'components/elements/Subhead'
 import Text from 'components/elements/Text'
-import { InteractiveExample } from 'pages/embed'
+import { InteractiveExample } from 'components/pages/embed'
 
 const Heading = withTitle(HeadingBase)
 const Subhead = withTitle(SubheadBase)
@@ -839,7 +839,8 @@ const Hero = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
         >
           Get Started Free
         </ArrowLink>
@@ -1216,7 +1217,8 @@ const CTASection = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
           css={theme({ fontSize: 3, px: 5, py: 3, color: 'white' })}
         >
           Start Building Free
@@ -1395,7 +1397,8 @@ const TryItSection = () => (
         <Link
           href='/docs/guides/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
           css={theme({ fontSize: CTA_LINK_FONT_SIZE })}
         >
           Start now for free
